move common backend definitions into backend/common
